% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_barplotdata.R
\name{hsrdef_barplotdata}
\alias{hsrdef_barplotdata}
\title{Function for reactive data update before plotting}
\usage{
hsrdef_barplotdata(x, data, code_levels)
}
\arguments{
\item{x}{the original SAS upload data}

\item{data}{the modified data with Include and Exclude labels}

\item{code_levels}{the code levels to include in the plots}
}
\value{
bar plot data
}
\description{
Function for reactive data update before plotting
}
