% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_body.R
\name{hsrdef_body}
\alias{hsrdef_body}
\title{Create main body of user interface for hsrdefbuilder app}
\usage{
hsrdef_body()
}
\value{
Input to dashboardBody
}
\description{
Create main body of user interface for hsrdefbuilder app
}
