% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_bodystyle.R
\name{hsrdef_bodystyle}
\alias{hsrdef_bodystyle}
\title{background colour html for user interface}
\usage{
hsrdef_bodystyle()
}
\value{
html style
}
\description{
background colour html for user interface
}
