% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_datatable.R
\name{hsrdef_datatable}
\alias{hsrdef_datatable}
\title{Build data table output in server}
\usage{
hsrdef_datatable(x)
}
\arguments{
\item{x}{reactive values to get data from}
}
\value{
data table to print
}
\description{
Build data table output in server
}
