% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_donutdata.R
\name{hsrdef_donutdata}
\alias{hsrdef_donutdata}
\title{Build reactive data set for donut chart or single bar chart}
\usage{
hsrdef_donutdata(df)
}
\arguments{
\item{df}{data set returned from bar chart hsrdef_table}
}
\value{
data set to use in second bar chart
}
\description{
Build reactive data set for donut chart or single bar chart
}
