% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_initialdata.R
\name{hsrdef_initialdata}
\alias{hsrdef_initialdata}
\title{Create inital data sets}
\usage{
hsrdef_initialdata(input_data_select)
}
\arguments{
\item{input_data_select}{input data from ui}

\item{values}{vals list from server}
}
\value{
values reactive list with updated data
}
\description{
Create inital data sets
}
