% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_sidebar.R
\name{hsrdef_sidebar}
\alias{hsrdef_sidebar}
\title{Create side-bar in shiny app ui}
\usage{
hsrdef_sidebar()
}
\value{
side-bar to include in user interface
}
\description{
Create side-bar in shiny app ui
}
