% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsrdef_uploaddata.R
\name{hsrdef_uploaddata}
\alias{hsrdef_uploaddata}
\title{Parse user uploaded file and add labels to data}
\usage{
hsrdef_uploaddata(file, old_data)
}
\arguments{
\item{file}{ui from upload handler}

\item{old_data}{current data table with code table}
}
\value{
updated data file
}
\description{
Parse user uploaded file and add labels to data
}
