% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function for hsrdefbuilder app}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{ui input}

\item{output}{plot and table outputs}

\item{session}{shiny server session}
}
\value{
A shiny server, to be called when running app
}
\description{
Server function for hsrdefbuilder app
}
